refactor: extract review bar component
